+++ app/MainWindow_SavedSets.cpp
@@ -10,34 +10,36 @@
 /// Related to the saving, loading and management of saved Armor Sets
 /// Also restoring backup
 
-void MainWindow::_loadSavedSet()
+void MainWindow::_populateSavedSets()
 {
-    if (_SavedSets.empty())
-    {
-        DialogWindow *Dia = new DialogWindow(this, "Warning", "There are no saved sets", Status::WARNING);
-        Dia->show();
-        return;
-    }
-
     QStringList items;
     for (const auto &it : _SavedSets.items())
         items << it.key().c_str();
+    
+    ui->savedComboBox->addItems(items);
+}
 
-    bool ok;
-    QString text = getItemInputDialog(this, "Select Armor Set", "Select set: ", items, &ok);
-    if (!ok)
+void MainWindow::_loadSavedSet()
+{
+    int currentIdx = ui->savedComboBox->currentIndex();
+
+    if (currentIdx == -1)
+    {
+        DialogWindow *Dia = new DialogWindow(this, "Warning", "No saved set selected", Status::WARNING);
+        Dia->show();
         return;
-
+    }
+    std::string currentText = ui->savedComboBox->currentText().toStdString();
     try
     {
         for (int i = 0; i < 5; ++i)
-            _MHManager.getPlayerData().setArmorPiece(i, _SavedSets[text.toStdString()][i]);
+            _MHManager.getPlayerData().setArmorPiece(i, _SavedSets[currentText][i]);
     }
     catch (std::exception &e)
     {
         DialogWindow *Dia = new DialogWindow(this, "ERROR", "Invalid Value for armor", Status::ERROR0);
         Dia->show();
-        DEBUG_LOG(ERROR, "Invalid value at json file. Error " << e.what());
+        DEBUG_LOG(ERROR, "Invalid value at saved sets json file. Error " << e.what());
         return;
     }
     this->_updateArmorValues();
@@ -48,38 +50,45 @@ void MainWindow::_saveCurrentSet()
     if (!this->_parseInputBoxes())
         return;
 
-    bool ok;
-    QString text;
-    while (true)
-    {
-        text = getTextInputDialog(this, "Saving Armor set", "Name for the set", &ok);
-        if (!ok)
-            return;
-
-        if (_SavedSets.find(text.toStdString()) != _SavedSets.end())
-        {
-            DialogWindow *Dia = new DialogWindow(this, "WARNING", "The name already exists.", Status::WARNING);
-            Dia->show();
-            while (Dia->isVisible())
-                QCoreApplication::processEvents();
-            continue;
-        }
-
-        if (!text.isEmpty())
-            break;
-    }
+    std::string currentText = ui->savedComboBox->currentText().toStdString();
+    if (_SavedSets.find(currentText) == _SavedSets.end())
+        ui->savedComboBox->insertItem(0, currentText.data());
 
     auto Data = _MHManager.getPlayerData().getData();
-    _SavedSets[text.toStdString()] = Data;
+    _SavedSets[currentText] = Data;
 
     if (!this->_flushSavedSets())
     {
-        DialogWindow *Dia = new DialogWindow(this, "ERROR", "Couldn't save set.", Status::ERROR0);
+        DialogWindow *Dia = new DialogWindow(this, "ERROR", "Couldn't save to file.", Status::ERROR0);
         Dia->show();
         return;
     }
     DialogWindow *Dia = new DialogWindow(this, "Success", "Successfully saved set!", Status::SUCCESS);
     Dia->show();
+}
+
+void MainWindow::_deleteCurrentSet()
+{
+    int currentIdx = ui->savedComboBox->currentIndex();
+
+    if (currentIdx == -1)
+    {
+        DialogWindow *Dia = new DialogWindow(this, "Warning", "No saved set selected", Status::WARNING);
+        Dia->show();
+        return;
+    }
+    std::string currentText = ui->savedComboBox->currentText().toStdString();
+    ui->savedComboBox->removeItem(currentIdx);
+    if (_SavedSets.find(currentText) == _SavedSets.end())
+        return;
+    _SavedSets.erase(currentText);
+    
+    if (!this->_flushSavedSets())
+    {
+        DialogWindow *Dia = new DialogWindow(this, "ERROR", "Couldn't save to file.", Status::ERROR0);
+        Dia->show();
+        return;
+    }
 }
 
 bool MainWindow::_flushSavedSets()